Backend/app/agents/media_agent.py
"""Media Agent — generates educational images and audio using MCP GenMedia.

Transferred to when the tutor wants to generate illustrations, diagrams,
or pronunciation audio for the student.

Uses the MCP genmedia server (Imagen 3 / Chirp 3 HD) via MCPToolset.
"""

from __future__ import annotations

import logging

from google.adk.agents import Agent
from app.tools.media_tools import MediaTools

from app.config import settings

logger = logging.getLogger(__name__)

MEDIA_AGENT_INSTRUCTION = """\
You are the **Media Agent** — the creative assistant inside the Magic
Whiteboard Tutor.

## Your responsibilities
- Generate educational illustrations, diagrams, and concept art using
  Imagen 3 (via the genmedia MCP server).
- Generate pronunciation audio clips using Chirp 3 HD when teaching
  languages or vocabulary.

## Guidelines
1. **Respect the student's request** — if the student (or tutor agent)
   asks for a specific subject (e.g. “a cat in space”), your image MUST
   depict that subject. Do not silently switch to a different topic like
   “the water cycle” unless the student explicitly changes the request
   or the tutor agent clearly asks for another concept.
2. **Educational focus** — when possible, make the image helpful for
   learning (clear composition, labels, age-appropriate), but never at
   the cost of changing the requested subject.
3. **Prompt engineering** — write detailed, specific prompts. Include
   phrases like “educational illustration”, “labelled diagram” or
   “whiteboard style” as appropriate, while keeping the topic exactly
   aligned with the student's request.
4. Tell the student what you're generating and give a brief description
   after it's ready, in plain language that matches what they asked for.
5. If the media generation service is unavailable or you cannot safely
   fulfil the request, explain why and transfer back to the tutor agent
   instead of substituting a different topic.

When media generation is complete, transfer back to the tutor agent.
"""


def build_media_agent() -> Agent:
    """Build the media agent for generating and showing images."""
    m_tools = MediaTools()
    return Agent(
        name="media_agent",
        model=settings.media_agent_model,
        instruction=MEDIA_AGENT_INSTRUCTION,
        tools=[m_tools.generate_and_show_image],
    )
